models in folder ok, POST method testing, joinLobby ok
