+++ srcs/app/transcendence/admin.py
@@ -6,6 +6,7 @@ from .models import *
 # 👇 2. Add this line to add the notification
 admin.site.register(Notification)
 admin.site.register(CustomUser)
+admin.site.register(FriendRequest)
 admin.site.register(Game)
 admin.site.register(Stat_Game)
 admin.site.register(Lobby)
