+++ srcs/app/transcendence/api.py
@@ -1,5 +1,6 @@
 from django.urls import path
 from .methods import *
+from .models import *
 
 urlpatterns = [
 	path('login/', login_user, name='login'),
@@ -11,14 +12,13 @@ urlpatterns = [
 	path('get_user_by_id/<int:id>', getUserById, name='getUserById'),
 	path('get_user_by_username/<str:username>', getUserByName, name='getUserByName'),
 	path('get_user_name/', getUserName, name='getUserName'),
+	path('me', getMe, name='getMe'),
 	path('get_user_connected/', getUserConnected, name='getUserConnected'),
 	path('get_game_by_id/<int:id>', getGameById, name='getGameById'),
 	path('get_game_by_name/<str:name>', getGameByName, name='getGameByName'),
 	path('get_all_games/', getAllGames, name='getAllGames'),
 	path('get_stats_users_by_game/<int:id_game>', getStatsUsersByGame, name='getStatsUsersByGame'),
-	path('add_win_to_user/<int:id_game>', addWinToUser, name='addWinToUser'),
-
-	path('get_user_history_by_game/<int:id_game>', getUserHistoryByGame, name='getUserHistoryByGame'),
+	
 	path('get_all_party/', getAllParties, name='getPartyById'),
 
 	path('get_all_lobbies/', getAllLobbies, name='getAllLobbies'),
@@ -31,5 +31,7 @@ urlpatterns = [
 	path('get_all_friends/<int:id_user>', getAllFriendsofUser, name='getAllFriendsofUser'),
 
 	path('join_lobby/', joinLobby, name='joinLobby'),
+	path('quit_lobby/', quitLobby, name='quitLobby'),
+	path('add_win_to_user/<int:id_game>', addWinToUser, name='addWinToUser'),
 
 ]
