+++ srcs/app/transcendence/methods/lobby.py
@@ -115,7 +115,6 @@ logger = logging.getLogger(__name__)
 @login_required
 @require_http_methods(['POST', 'PUT'])
 def joinLobby(request):
-	logger.error('joinLobbyHEHEHEHEHEHEHEHHEHEHE')
 	data = json.loads(request.body)
 	id_game = data['id_game']
 	type = data['type'] if 'type' in data else 'Public' #todo: check if type is valid
@@ -124,20 +123,18 @@ def joinLobby(request):
 		lobby = game.lobby_set.all()
 		#if no lobby for this game, create a new lobby
 		if len(lobby) == 0:
-			logger.error('no lobby for this game')
 			lobby = newLobby(game, type)
 		#if there is a lobby of given type, join the lobby
 		else:
 			lobby = lobby.filter(type=type)
-			logger.error('lobby of given type')
 			if len(lobby) == 0:
 				lobby = newLobby(game, type)
 			else:
 				lobby = lobby[0]
-		logger.error('lobby created')
 		user = request.user
-		userInLobby = UserInLobby.objects.create(id_user=user, id_lobby=lobby)
-		Lobby.objects.get(id=lobby.id).user.add(userInLobby)
+		if (UserInLobby.objects.filter(id_user=user, id_lobby=lobby).count() > 0):
+			return JsonResponse({'status': 'ok', 'id_lobby': lobby.id})
+		UserInLobby.objects.create(id_user=user, id_lobby=lobby)
 		return JsonResponse({'status': 'ok', 'id_lobby': lobby.id})
 	except Game.DoesNotExist:
 		return JsonResponse({'status': 'error', 'message': 'This game does not exist.'}, status=404)
@@ -154,14 +151,14 @@ def quitLobby(request):
 	id_lobby = data['id_lobby']
 	try:
 		lobby = Lobby.objects.get(id=id_lobby)
-		userInLobby = UserInLobby.objects.get(id_user=request.user, id_lobby=lobby)
-		Lobby.objects.get(id=lobby.id).user.remove(userInLobby)
-		userInLobby.delete()
-		return JsonResponse({'status': 'ok', 'message': 'You have left the lobby.'})
+		userInLobby = UserInLobby.objects.filter(id_user=request.user, id_lobby=lobby).first()
+		if userInLobby:
+			userInLobby.delete()
+			return JsonResponse({'status': 'ok', 'message': 'You have left the lobby.'})
+		else:
+			return JsonResponse({'status': 'error', 'message': 'You are not in this lobby.'}, status=400)
 	except Lobby.DoesNotExist:
 		return JsonResponse({'status': 'error', 'message': 'This lobby does not exist.'}, status=404)
-	except UserInLobby.DoesNotExist:
-		return JsonResponse({'status': 'error', 'message': 'You are not in this lobby.'}, status=400)
