+++ srcs/app/transcendence/methods/party.py
@@ -35,18 +35,6 @@ def getAllParties(request):
 
 @login_required
 @require_http_methods(['GET'])
-def getUserHistoryByGame(request, id_game):
-    try:
-        game = Game.objects.get(id=id_game)
-        user = request.user
-        parties = Party.objects.filter(id_game=game, player1=user) | Party.objects.filter(id_game=game, player2=user)
-        data = [party.party_data() for party in parties]
-        return JsonResponse({'status': 'ok', 'parties': data})
-    except Game.DoesNotExist:
-        return JsonResponse({'status': 'error', 'message': 'This game does not exist.'}, status=404)
-
-@login_required
-@require_http_methods(['GET'])
 def getPartyByGame(request, id_game):
 	try:
 		game = Game.objects.get(id=id_game)
