+++ srcs/app/transcendence/models/__init__.py
@@ -1,0 +1,7 @@
+from .CustomUser import CustomUser, FriendRequest
+from .Notification import Notification
+from .Stat import  Stat_User_by_Game
+from .Game import Game, Stat_Game
+from .Lobby import Lobby, UserInLobby
+from .Party import Party
+from .Tournament import Tournament
